requirements.txt
allure-pytest==2.12.0
PyYAML==6.0
pytest==7.2.0
selenium==4.6.0
webdriver-manager==3.8.5
pytest-xdist==3.2.0
